Stage 1: Extract constants + compat.py scaffolding
