+++ python/sglang/srt/arg_groups/__init__.py
@@ -1,0 +1,44 @@
+"""Shared server / CLI argument groups (choice lists, helpers).
+
+Public names re-exported via ``sglang.srt.server_args`` are listed in ``__all__``.
+"""
+
+__all__ = [
+    "ATTENTION_BACKEND_CHOICES",
+    "DEFAULT_LORA_EVICTION_POLICY",
+    "DEFAULT_UVICORN_ACCESS_LOG_EXCLUDE_PREFIXES",
+    "DETERMINISTIC_ATTENTION_BACKEND_CHOICES",
+    "DISAGG_TRANSFER_BACKEND_CHOICES",
+    "ENCODER_TRANSFER_BACKEND_CHOICES",
+    "FP4_GEMM_RUNNER_BACKEND_CHOICES",
+    "FP8_GEMM_RUNNER_BACKEND_CHOICES",
+    "GRAMMAR_BACKEND_CHOICES",
+    "LINEAR_ATTN_KERNEL_BACKEND_CHOICES",
+    "LOAD_FORMAT_CHOICES",
+    "LORA_BACKEND_CHOICES",
+    "MAMBA_BACKEND_CHOICES",
+    "MAMBA_SCHEDULER_STRATEGY_CHOICES",
+    "MOE_A2A_BACKEND_CHOICES",
+    "MOE_RUNNER_BACKEND_CHOICES",
+    "NSA_CHOICES",
+    "NSA_PREFILL_CP_SPLIT_CHOICES",
+    "PREFILL_CP_SPLIT_CHOICES",
+    "QUANTIZATION_CHOICES",
+    "RADIX_EVICTION_POLICY_CHOICES",
+    "RADIX_SUPPORTED_DETERMINISTIC_ATTENTION_BACKEND",
+    "RL_ON_POLICY_TARGET_CHOICES",
+    "SAMPLING_BACKEND_CHOICES",
+    "SPECULATIVE_DRAFT_MODEL_QUANTIZATION_CHOICES",
+    "add_attention_backend_choices",
+    "add_deterministic_attention_backend_choices",
+    "add_disagg_transfer_backend_choices",
+    "add_fp4_gemm_runner_backend_choices",
+    "add_fp8_gemm_runner_backend_choices",
+    "add_grammar_backend_choices",
+    "add_load_format_choices",
+    "add_moe_runner_backend_choices",
+    "add_quantization_method_choices",
+    "add_radix_eviction_policy_choices",
+    "add_radix_supported_deterministic_attention_backend_choices",
+    "add_rl_on_policy_target_choices",
+]
